- generate distinct nbest surface strings (if 3rd token of n-best-list is 'distinct')  - some sanity checks
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk/moses@674 1f5c12ca-751b-0410-a591-d2e778427230

+++ src/PhraseDictionaryTreeAdaptor.cpp
@@ -459,7 +459,7 @@ struct PDTAimp {
 			{
 				assert(i->first.first<m_rangeCache.size());
 				assert(i->first.second>0);
-				assert(i->first.second-1<m_rangeCache[i->first.first].size());
+				assert(static_cast<size_t>(i->first.second-1)<m_rangeCache[i->first.first].size());
 				assert(m_rangeCache[i->first.first][i->first.second-1]==0);
 
 				std::vector<TargetPhrase> tCands;tCands.reserve(i->second.size());
